requirements.txt
Flask==2.0.1
Flask-SQLAlchemy==2.5.1
Flask-Migrate==3.0.1
python-dotenv==0.17.0
pytest==6.2.4
pytest-cov==2.12.1
